Amended the copy for full remission confirmation
Modified the copy change for the full remission confirmation page.

Also tried to make the tests for it clear.
